refactored tissue expression view
